bandmate/requirements.txt
flask==3.0.3
gunicorn==22.0.0
pillow==10.4.0
pytesseract==0.3.10
python-dotenv==1.0.1
openai==1.40.0
python-docx==1.1.2
reportlab==4.2.2
